--- app/views/author/posts/new.html.erb
@@ -1,3 +1,0 @@
-<h1>New Post</h1>
-<%= render 'form', post: @post %>
-<%= link_to 'Back', posts_path, class: "btn btn-primary" %>
