+++ app/views/authors/posts/_edit_form.html.erb
@@ -1,4 +1,4 @@
-<%= form_with(model: post, local: true, url: author_posts_path, remote: true) do |form| %>
+<%= form_with(model: post, local: true, url: authors_post_path, remote: true) do |form| %>
   <% if post.errors.any? %>
     <div id="error_explanation">
       <h2><%= pluralize(post.errors.count, "error") %> prohibited this post from being saved:</h2>
