+++ app/views/authors/posts/edit.html.erb
@@ -1,4 +1,4 @@
 <h1>Editing Post</h1>
 <%= render 'edit_form', post: @post %>
 <%= link_to 'Show', @post %> |
-<%= link_to 'Back', posts_path %>
+<%= link_to 'Back', authors_posts_path %>
