+++ app/views/authors/posts/new.html.erb
@@ -1,0 +1,3 @@
+<h1>New Post</h1>
+<%= render 'form', post: @post %>
+<%= link_to 'Back', authors_posts_path, class: "btn btn-primary" %>
