+++ app/views/shared/_navbar.html.erb
@@ -13,7 +13,7 @@
             <%= current_user.name %>
           </a>
           <div class="dropdown-menu" aria-labelledby="navbarDropdown">
-            <%= link_to 'My posts', author_posts_path, class: "dropdown-item" %>
+            <%= link_to 'My posts', authors_posts_path, class: "dropdown-item" %>
             <%= link_to 'Edit Account', edit_user_path(current_user), class: "dropdown-item" %>
             <div class="dropdown-divider"></div>
              <%= link_to 'Log out', logout_path, class: "dropdown-item"  %>
